Move Run method from TestRunner to TestBase, remove TestRunner
