+++ Tests/118_EnhancedStrongName.Test/EnhancedStrongNameTest.cs
@@ -13,7 +13,7 @@ namespace EnhancedStrongName.Test {
 		[Trait("Category", "core")]
 		[Trait("Issue", "https://github.com/mkaring/ConfuserEx/issues/118")]
 		public async Task EnhancedStrongName() =>
-			await TestRunner.Run("118_EnhancedStrongName.exe",
+			await Run("118_EnhancedStrongName.exe",
 				new[] {"My strong key token: 79A18AF4CEA8A9BD", "My signature is valid!"},
 				null,
 				outputHelper,
